Show the right sidebar expanded by default
Today, users are completely missing the sidebar, this is a big problem
cos part of crucial BR Modelo Web functionalities lives in the sidebar.

As a quick solution, we'll show the right sidebar expanded by default,
and add an empty state block to be visible when the user has no element
selected.