Refactor login tests and custom command
The login tests now have better assertions.

And the login custom command can now use the cy.session
experimental feature, which allows restoring the session of a user,
which speeds up tests.

But, for the login tests, we don't want any cache to be restored.
This is why for the login tests we pass { cacheSession: false }