Update createModelViaApi custom command
So that we can pass a model definition (defined as fixture) to
easily create all kinds of models without having to go through the
graphical user interface (GUI).

If no model is passed, an empty one is used as a default.

The custom command type definition has been upated accordingly.