Make test more robust
When running this test against a remote MongoDB, things happen a
bit slower than when running it against a local db.

Without this change, the `cy.reload()` command can happen too fast,
making the POST request to `/models` to be canceled.

With this change, we make sure `cy.reload()` only happens after
Cypress has waited for the request to happen.