Change bug report icon position
Based in users feedback, we realized that keep the "bug" icon in
the bottom right portion of the page may sometimes cause a
problem  when the icon stays on top of .model-properties content

This happens specially when a table or entity with larger portions
of information is selected.

To avoid this I'm moving the "report bug" to the botom right
portion of the page.